Marks Verification WorkFlow using SpringWebFlow
The project consists in developing a web application to simplify the
marks verification procedure after a student request. The student
fulills a request online and the administration will handle it then
following a predefined workflow. Users of the application are students,
professors and administration employees. Used Technologies : Spring
WebFlow - Spring Security